Add Accuracy property to GameAttempt entity
This commit introduces a new `Accuracy` property of type `decimal` to the `GameAttempt` entity in the database. The property is added to the `GameAttempts` table, marked as required with a default value of `0m`. Updates are made to the `GameAttempt` class to include this property, and the `AccessorDbContextModelSnapshot` is modified to reflect the current state of the database schema.